chain: Move owasm to pkg folder
