+++ chain/app/genesis.go
@@ -175,84 +175,84 @@ func GetDefaultDataSourcesAndOracleScripts(owner sdk.AccAddress) json.RawMessage
 		{
 			"Crypto price script (Borsh version)",
 			"Oracle script for getting an average crypto price from many sources encoding parameter by borsh.",
-			"./owasm/res/crypto_price_borsh.wasm",
+			"./pkg/owasm/res/crypto_price_borsh.wasm",
 			`{"Input": "{\"kind\": \"struct\", \"fields\": [ [\"symbol\", \"string\"], [\"multiplier\", \"u64\"] ] }","Output": "{ \"kind\": \"struct\", \"fields\": [ [\"px\", \"u64\"] ]}"}`,
 			`https://ipfs.io/ipfs/QmUrYgDKXT8V8DPdCYMEwPM6n82r6zxbvBf6p4gb4m1RA5`,
 		},
 		{
 			"Gold price script",
 			"Oracle script for getting an average gold price in ATOM",
-			"./owasm/res/gold_price.wasm",
+			"./pkg/owasm/res/gold_price.wasm",
 			`{"Input": "{\"kind\": \"struct\", \"fields\": [ [\"multiplier\", \"u64\"] ] }","Output": "{ \"kind\": \"struct\", \"fields\": [ [\"px\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmPheBfYjM4fZ6ngSHYrnDgmapZi9r1i4x5hGFUUyZiP5y`,
 		},
 		{
 			"Alphavantage stock price script",
 			"Oracle script for getting an average stock price from Alphavantage",
-			"./owasm/res/alphavantage.wasm",
+			"./pkg/owasm/res/alphavantage.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"symbol\", \"string\"],[\"api_key\", \"string\"], [\"multiplier\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"px\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmSnkymmSw4Ho46xazbJ3CJ5kp2hw2ZUfqNc7GJzuv7GCZ`,
 		},
 		{
 			"Bitcoin block count",
 			"Oracle script for getting Bitcoin latest block height",
-			"./owasm/res/bitcoin_block_count.wasm",
+			"./pkg/owasm/res/bitcoin_block_count.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"_unused\", \"u8\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"block_count\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmZpccMgt1gbcmv1ahK727xUpH1NHex3J22kQua1SNGVkN`,
 		},
 		{
 			"Bitcoin block hash",
 			"Oracle script for getting Bitcoin latest block hash",
-			"./owasm/res/bitcoin_block_hash.wasm",
+			"./pkg/owasm/res/bitcoin_block_hash.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"block_height\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"block_hash\", \"string\"] ] }"}`,
 			`https://ipfs.io/ipfs/Qmcgv1vHLVNQPKkkAB9ftG33142E8Ufbm9GUpGznbh5NzW`,
 		},
 		{
 			"Coingecko crypto volume",
 			"Oracle script for getting an average crypto price from Coingecko",
-			"./owasm/res/coingecko_volume.wasm",
+			"./pkg/owasm/res/coingecko_volume.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"symbol\", \"string\"], [\"multiplier\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"volume\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmNSKWfAgairPCKjypZsueStztrP1tBmtWEukmi2Udejx2`,
 		},
 		{
 			"Crypto compare crypto volume",
 			"Oracle script for getting an average crypto price from Crypto compare",
-			"./owasm/res/crypto_compare_volume.wasm",
+			"./pkg/owasm/res/crypto_compare_volume.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"symbol\", \"string\"], [\"multiplier\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"volume\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmQD5bvn8esMj8ajJYii6U6d34g4b9vaNJRG5zw42Yf1NC`,
 		},
 		{
 			"Ethereum gas price",
 			"Oracle script for getting gas price from ETH gas station",
-			"./owasm/res/eth_gas_station.wasm",
+			"./pkg/owasm/res/eth_gas_station.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"gas_option\", \"string\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"gweix10\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmQjSzHx4YkfdaSs7uAnJu9qhCXZtTKRxv1mn3v7Z27hS6`,
 		},
 		{
 			"Open sky network",
 			"Oracle script for getting the verification of a flight",
-			"./owasm/res/open_sky_network.wasm",
+			"./pkg/owasm/res/open_sky_network.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"flight_op\", \"string\"],[\"airport\", \"string\"],[\"icao24\", \"string\"],[\"begin\", \"string\"], [\"end\", \"string\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"flight_existence\", \"u8\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmTGWmf69iQp8XecRcxvzxXmRXyfNTgscudtwRH7pLy8TF`,
 		},
 		{
 			"Open weather map",
 			"Oracle script for getting weather information",
-			"./owasm/res/open_weather_map.wasm",
+			"./pkg/owasm/res/open_weather_map.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"country\", \"string\"],[\"main_field\", \"string\"],[\"sub_field\", \"string\"],[\"multiplier\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"value\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmbJ9csujaPFdY58aXjK7vfSRYLAhN3rNH1vSPdUF4mXyv`,
 		},
 		{
 			"Quantum random number generator",
 			"Oracle script for getting a big random number from quantum computer",
-			"./owasm/res/qrng.wasm",
+			"./pkg/owasm/res/qrng.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"size\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"random_bytes\", \"string\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmbUCBoWvKiHGktpdhUPMEm6hvQnM7xgWfUm2HVDfwcCgS`,
 		},
 		{
 			"Yahoo stock price",
 			"Oracle script for getting stock price from Yahoo",
-			"./owasm/res/yahoo_price.wasm",
+			"./pkg/owasm/res/yahoo_price.wasm",
 			`{"Input":"{ \"kind\": \"struct\", \"fields\": [ [\"symbol\", \"string\"], [\"multiplier\", \"u64\"] ] }","Output":"{ \"kind\": \"struct\", \"fields\": [ [\"px\", \"u64\"] ] }"}`,
 			`https://ipfs.io/ipfs/QmbgUQq82ra3bnxu8Jg89uonKQFpuiXx1xecXUBun2AcjF`,
 		},
