+++ chain/pkg/owasm/utils.go
