+++ chain/run_tests.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
+mkdir -p pkg/owasm/res
+
 # release mode
 cd ../owasm/chaintests
 
 for f in *; do
     if [ -d "$f" ]; then
         RUSTFLAGS='-C link-arg=-s' cargo build --target wasm32-unknown-unknown --release --package $f
-        cp ../target/wasm32-unknown-unknown/release/$f.wasm ../../chain/owasm/res
+        cp ../target/wasm32-unknown-unknown/release/$f.wasm ../../chain/pkg/owasm/res
     fi
 done
 cd ../../chain
@@ -16,7 +18,7 @@ cd ../owasm/chaindebugtests
 for f in *; do
     if [ -d "$f" ]; then
         RUSTFLAGS='-C link-arg=-s' cargo build --target wasm32-unknown-unknown --package $f
-        cp ../target/wasm32-unknown-unknown/debug/$f.wasm ../../chain/owasm/res
+        cp ../target/wasm32-unknown-unknown/debug/$f.wasm ../../chain/pkg/owasm/res
     fi
 done
